Audio: Use a QVarLengthArray instead of a variable-size stack buffer
"QVarLengthArray allocates a certain number of elements on the stack,
and if you resize the array to a larger size, it automatically uses
the heap instead."

In older versions of GCC (<4.9), initialization of the buffer with the
braces would not work. In any case, variable length stack buffers was
not the best idea here since we don't know how big it will have to be.