Core: Use QByteArray to store system_directory
In class Core, storing strings as QStrings internally is a bad idea
since they use utf-16 internally.

We would have to convert strings to ASCII/utf-8 every time we
talk with the core.

Moreover, simply doing myQStr.toStdString().c_str() to convert them
is a *bad idea*. As soon as the std::string goes out of scope, the
buffer will be freed. If the c_str() pointer is returned to the core,
undefined behavior will ensue.