Changed Core to a QObject (read description)
You have to clean and maybe even nuke the build directory in order to
get it to build. QMake doesn’t handle changing a normal C++ class to a
QObject, apparently.

I changed the connect syntax in VideoItem to the newer style. Need to
do this for the rest of the codebase.

I also added some delete calls to Audio and got rid of some redundant
unload() functions.