MarqueeText: Added new "forward" property, reverse mode
Setting this to false (reverse mode) will reverse the scroll direction,
right justify and left elide:
“This is a line of text”
forward = true: “This is a…”
forward = false: “…of text”

In some cases the text may move a bit when going to its idle position
in reverse mode. I’ll look into that someday, maybe.