Make use of native fullscreen hotkey
This commit allows the user to press the native hotkey on their
system to fullscreen the application. This is based on desktop
environment and so is `CTRL+F11` on GNOME and `F11` on most other
platforms.

This works well because some systems (such as chromebooks) do not
have an F11 key and must make use of other system defined hotkeys

Changes:

	1. Added `id:` of fullscreenItem to the Item that handles
	   the fullscreen game action

	2. Added contents of the MouseArea `onClicked:` signal
	   to a function inside fullscreenItem called
	   prepareFullscreen()

	3. Added a shortcut to StandardKey.Fullscreen which calls
	   prepareFullscreen() as its callback